Fixing float issues

+++ lesson2/solution/lib/src/main/java/com/example/Animal.java
@@ -29,7 +29,7 @@ public abstract class Animal {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
